requirements.txt
python-dotenv
flask
openai==0.28
requests
streamlit
beautifulsoup4